Restructure code to make the code more OOP
