+++ src/main/java/org/ajay/common/Constants.java
@@ -1,0 +1,11 @@
+package org.ajay.common;
+
+public class Constants {
+    // Constants
+    public static final String CHAT_BOT_NAME = "Jarvis"; // Name of the chatbot
+
+    public static final String FILE_PATH = "data/jarvis.txt"; // File path to store the tasks
+
+    // Constants for the commands
+    public final static String EXIT_COMMAND_ALT = "bye";
+}
