--- src/main/java/org/ajay/utils/Constants.java
@@ -1,7 +1,0 @@
-package org.ajay.utils;
-
-public class Constants {
-    // Constants
-    public static final String chatBotName = "Jarvis"; // Name of the chatbot
-
-}
